P7-T02: the load run, and the five defects it found
The acceptance asks for hundreds of concurrent members in one
workspace. This machine holds every §13.1 budget at twenty, up from
ten when the row was written, and what moved it was five measured
defects rather than tuning.

**A drag locked the whole column.** `moveTaskInTx` read the
destination column under `select ... for update` to find the two
positions a drop sits between. A column in §13.1's dataset holds
18,937 cards, so one drag row-locked nineteen thousand rows and held
them until commit: 20,394ms at the 95th percentile while every read
stayed under two seconds. It never needed the column, only the card it
landed after and the one after that, and the advisory lock over the
space's column already serialises drags. Three `limit 1` queries:
2,087ms.

**Then the renumber underneath it.** Whole-number positions are spaced
1024 apart and each drop into the same slot halves the gap, so after
about ten the column is rewritten evenly, nineteen thousand rows in
the same transaction. That was the fifteen-to-twenty-second tail on
one drag in ten, and the evidence is in the data rather than the
reasoning: 187,285 task rows carried an `ordering_state.normalisedAt`
stamp. Agung chose fractional positions on 10 September 2026.
Migration 0081 makes `position` `double precision`, so the same gap
takes about forty halvings, and a drop at the top of a column now goes
*below* the first card rather than half way to zero: nothing requires
a position to be positive, and dropping at the top is the commonest
drag there is. `double precision` and not `numeric` because
node-postgres returns a numeric as a string.

**The access filter was building a million rows to answer a page.**
Postgres rewrites an `EXISTS` whose only tie to the outer row is one
equality into a hashed subplan: it builds the inner set once and
probes it. For `accessScopeFilter` that set is every context the
reader can see, 1,271,774 rows on this dataset, built to answer a
fifty-row feed page. 468ms of the query's 540, and it does not shrink
when the page does. `offset 0` fences the rewrite off and leaves one
index lookup per candidate row: the page costs 2ms and throughput over
the whole run went from 29.4 to 61.7 calls a second. The two facts the
filter used to ask the database per row, live and human, now come in
through `accessFilterMember`, because neither depends on the outer row
and hoisting them is what exposed the rewrite.

**Publishing one check-in recomputed ten thousand goals.** The scope
of a recompute was the changed goal's whole cycle, and every row in it
was written one statement at a time: 20,034 statements, 8.9 seconds,
and at twenty concurrent members 47 seconds at the median with
nineteen deadlocks in thirty calls. Agung chose narrowing the scope on
10 September 2026 over deferring the recompute to a job. Progress only
rolls upward, so a change to one goal loads that goal, the goals above
it, and the siblings each of those rolls up with; a sibling arrives as
a settled boundary carrying the progress it already has, because its
own subtree is not loaded and recomputing it from half a graph would
read it as lower than it is. A cycle scope still means the cycle. The
writes are one statement per table and only for rows whose derived
values moved, which also stops `updated_at` being bumped on ten
thousand goals nobody touched. 17ms.

**And the dataset was lying again, for the third time.** Every
`parent_goal_id` was null, so the alignment score, the cascade and the
recompute were all measured over a hundred thousand roots. The seeder
now aligns each cycle's goals into a tree of fanout five, which is six
levels deep at ten thousand goals.

**Measured, on the full dataset, 10 September 2026.** At twenty
members every scenario is inside its own §13.1 ceiling with no errors,
and a fifteen-minute soak ran 40,987 calls with no errors and a drift
of 1.04x. Twenty-five holds everything with the drag on the line.
Thirty-five and fifty are the processor rather than any one query.

The harness gained what the measurement needed: `--pool`, because with
the audit lock gone every scenario settled into one band at fifty
members, which is a queue for connections and not a slow query; and
two scenarios. `reorder` drops a card into a real slot, because the
`drag` scenario never names a card to land after and the renumber path
would have gone unmeasured. `check-in` is the heaviest write there is
and the first thing this task's deliverables name, and it is what
found the recompute.

**Not covered here, stated rather than implied.** Realtime fan-out to
twenty session participants, chat inbound, external agent traffic and
channel nudge delivery all need a running server or a provider, and
§13.1 hands the same rows to the same places. This is not an HTTP load
test: Next.js and the reverse proxy in front of `callAction` are not
exercised.

Verified: typecheck, lint, dead-code, db:lint, boundaries, contract,
method, licences, and `recompute-scope.test.ts` (6 tests, including the
mutation that proves its own check can fail). Not run here, and owed before this is pushed: the full `packages/core` and `apps/web` suites,
the end-to-end suite, sign-off, and the Helm and Docker scripts this
machine has no tooling for.

+++ packages/core/src/actions/tasks.ts
@@ -78,7 +78,8 @@ const taskCard = z.object({
   title: z.string(),
   status: z.enum(TASK_STATUSES),
   dueOn: z.string().nullable(),
-  position: z.number().int(),
+  /** Fractional since migration 0081, so a drag writes one row. */
+  position: z.number(),
   assignees: z.array(assignee),
   checklist: z.object({ done: z.number().int(), total: z.number().int() }),
 });
@@ -920,7 +921,7 @@ export const moveTask = defineWriteAction({
   }),
   output: z.object({
     id: z.uuid(),
-    position: z.number().int(),
+    position: z.number(),
     /** True when the column was renumbered to make room. */
     normalised: z.boolean(),
   }),
